rustdoc: factor all-impls-for-item out into its own method
